README formatting
Okay, I found the right combination of formatting to do what I want. If
you want a second paragraph inside of a definition, you have to put an
unindented plus (+) and then unindent the second paragraph.

For example:

blah blah::
    Some definition text goes here.  This is paragraph one.
+
And now we continue on to the second paragraph. Note we drop the indent
to stop it from thinking we're a literal.