Update nagios-cli
This update brings a functional nagios-cli with built-in help. This lets
you easily schedule and cancel downtimes as well as output host/service
lists. More functionality will come as I add things to the API, but I
wanted to get something useful up.

The original JSON mode is still available, it's just buried with a --raw
option.