Add a few more essential_keys to return
These are required for my upcoming NagDash project which replaces the old Nagite2. (http://github.com/lozzd/Naglite2) 
Part of the information we want to display is how long a service has been in this state (thus last_state_change) 
and it also displays things that are about to alert (e.g. soft) so we need the number of attempts so far and the maximum attempts (which are also displayed)